requirements.txt
# Core API Requirements
flask==2.2.5
gunicorn==21.2.0
tensorflow-cpu==2.15.0
pillow==10.3.0
numpy==1.26.4
requests==2.31.0
python-dotenv==1.0.0

# Streamlit App Requirements
streamlit==1.13.0